Switch antlr to python 3 library, update usages
I got the antlr3 runtime from https://github.com/antlr/antlr3/tree/master/runtime/Python3 ,
then worked through a few failures. I may still need to regenerate the
parsers/lexers to get things to actually work at runtime, but at least things
import correctly now.

+++ google/appengine/api/datastore_types.py
@@ -780,7 +780,7 @@ class Key(object):
     return hash(type(args)) ^ hash(tuple(args))
 
 
-class _OverflowDateTime(long):
+class _OverflowDateTime(int):
   """Container for GD_WHEN values that don't fit into a datetime.datetime.
 
   This class only exists to safely round-trip GD_WHEN values that are too large
@@ -1097,7 +1097,7 @@ class PostalAddress(str):
     return '<gd:postalAddress>%s</gd:postalAddress>' % saxutils.escape(self)
 
 
-class Rating(long):
+class Rating(int):
   """A user-provided integer rating for a piece of content. Normalized to a
   0-100 scale.
